Common DraggablePolicyCard
